+++ lib/shared.ts
@@ -14,13 +14,6 @@ export const DEFAULT_PAYLOAD_LIMIT: number = 16777216;
 
 export const native: any = ((): NodeRequire => {
   try {
-    const [major, minor]: string[] = process.version.replace('v', '').split('.');
-
-    if (Number(major) === 13 && Number(minor) < 9) {
-      // temporary fix for node 13.8 and lower
-      return require(`../dist/bindings/cws_${process.platform}_${process.versions.modules}_8`);
-    }
-
     return require(`../dist/bindings/cws_${process.platform}_${process.versions.modules}`);
   } catch (err) {
     err.message = err.message + ` check './node_modules/@clusterws/cws/build_log.txt' for post install build logs`;
@@ -68,15 +61,7 @@ export function setupNative(group: any, type: string, wsServer?: WebSocketServer
     (webSocket as any).external = null;
 
     process.nextTick((): void => {
-      if (!code) {
-        // if no code provided it is 100% error in parsing or in code
-        webSocket.registeredEvents['error']({
-          message: 'cWs invalid status code or invalid UTF-8 sequence',
-          stack: 'cWs invalid status code or invalid UTF-8 sequence'
-        });
-      }
-
-      webSocket.registeredEvents['close'](code || 1006, message || '');
+      webSocket.registeredEvents['close'](code || 1005, message || '');
     });
 
     native.clearUserData(newExternal);