+++ src/headers/12/node_platform.h
@@ -78,15 +78,15 @@ class PerIsolatePlatformData :
   // posted during flushing of the queue are postponed until the next
   // flushing.
   bool FlushForegroundTasksInternal();
-  void CancelPendingDelayedTasks();
 
   const uv_loop_t* event_loop() const { return loop_; }
 
  private:
   void DeleteFromScheduledTasks(DelayedTask* task);
+  void DecreaseHandleCount();
 
   static void FlushTasks(uv_async_t* handle);
-  static void RunForegroundTask(std::unique_ptr<v8::Task> task);
+  void RunForegroundTask(std::unique_ptr<v8::Task> task);
   static void RunForegroundTask(uv_timer_t* timer);
 
   struct ShutdownCallback {
@@ -95,14 +95,18 @@ class PerIsolatePlatformData :
   };
   typedef std::vector<ShutdownCallback> ShutdownCbList;
   ShutdownCbList shutdown_callbacks_;
+  // shared_ptr to self to keep this object alive during shutdown.
+  std::shared_ptr<PerIsolatePlatformData> self_reference_;
+  uint32_t uv_handle_count_ = 1;  // 1 = flush_tasks_
 
+  v8::Isolate* const isolate_;
   uv_loop_t* const loop_;
   uv_async_t* flush_tasks_ = nullptr;
   TaskQueue<v8::Task> foreground_tasks_;
   TaskQueue<DelayedTask> foreground_delayed_tasks_;
 
   // Use a custom deleter because libuv needs to close the handle first.
-  typedef std::unique_ptr<DelayedTask, std::function<void(DelayedTask*)>>
+  typedef std::unique_ptr<DelayedTask, void(*)(DelayedTask*)>
       DelayedTaskPointer;
   std::vector<DelayedTaskPointer> scheduled_delayed_tasks_;
 };
@@ -137,7 +141,6 @@ class NodePlatform : public MultiIsolatePlatform {
   ~NodePlatform() override = default;
 
   void DrainTasks(v8::Isolate* isolate) override;
-  void CancelPendingDelayedTasks(v8::Isolate* isolate) override;
   void Shutdown();
 
   // v8::Platform implementation.
@@ -166,6 +169,8 @@ class NodePlatform : public MultiIsolatePlatform {
 
   std::shared_ptr<v8::TaskRunner> GetForegroundTaskRunner(
       v8::Isolate* isolate) override;
+
+  Platform::StackTracePrinter GetStackTracePrinter() override;
 
  private:
   std::shared_ptr<PerIsolatePlatformData> ForIsolate(v8::Isolate* isolate);